Integrated wheelspeeds into main autorally_controller for more accurate simulation
